new updates in routes in roles in permissions

+++ app/Http/Controllers/AdminController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Redirect;
 use App\Http\Requests\ServiceEditRequest;
 use App\Repositories\Contracts\AvisInterface;
 use App\Repositories\Contracts\ClientInterface;
@@ -16,9 +17,11 @@ use App\Repositories\Repository\ClientRepository;
 use App\Repositories\Contracts\CategorieInterface;
 use App\Repositories\Repository\ServiceRepository;
 use App\Repositories\Contracts\PrestataireInterface;
+use App\Repositories\Contracts\ReservationInterface;
 use App\Repositories\Contracts\UtilisateurInterface;
 use App\Repositories\Repository\CategorieRepository;
 use App\Repositories\Repository\PrestataireRepository;
+use App\Repositories\Repository\ReservationRepository;
 use App\Repositories\Repository\UtilisateurRepository;
 
 class AdminController extends Controller
@@ -29,8 +32,8 @@ class AdminController extends Controller
     private $UtilisateurRepository;
     private $ClientRepository;
     private $PrestataireRepository;
-
-    public function __construct(CategorieInterface $CategorieRepository, ServiceInterface $ServiceRepository,AvisInterface $AvisRepository,UtilisateurInterface $UtilisateurRepository,ClientInterface $ClientRepository,PrestataireInterface $PrestataireRepository)
+    private $ReservationRepository;
+    public function __construct(CategorieInterface $CategorieRepository, ServiceInterface $ServiceRepository,AvisInterface $AvisRepository,UtilisateurInterface $UtilisateurRepository,ClientInterface $ClientRepository,PrestataireInterface $PrestataireRepository,ReservationInterface $ReservationRepository)
     {
         $this->CategorieRepository =  $CategorieRepository;
         $this->ServiceRepository =  $ServiceRepository;
@@ -38,16 +41,37 @@ class AdminController extends Controller
         $this->UtilisateurRepository = $UtilisateurRepository;
         $this->ClientRepository = $ClientRepository;
         $this->PrestataireRepository = $PrestataireRepository;
+        $this->ReservationRepository = $ReservationRepository;
+    }
+
+    public function IndexHome()
+    {
+        $statisticusers = $this->UtilisateurRepository->GetStatisticUsers();
+        $lastusers = $this->UtilisateurRepository->GetLastUsers();
+        $reservations = $this->ReservationRepository->GetLastReservation();
+        return view('Admin.Dashboard-home',compact('statisticusers','lastusers','reservations'));
     }
 
     public function SelectCategories()
     {
+        if(!Auth::user()->HasPermission('create_category'))
+        {
+            abort(403);
+        }   
+        $StatisticCategrie = $this->CategorieRepository->GetStatisticCategorie();
         $categories = $this->CategorieRepository->ReadCategoriesPaginate();
-        return view('Admin.Dashboard-categorie', compact('categories'));
+        $service = $this->ServiceRepository->CountServices();
+
+
+        return view('Admin.Dashboard-categorie', compact('categories','StatisticCategrie','service'));
     }
 
     public function CreateCategorie(Request $request)
     {
+        if(!Auth::user()->HasPermission('create_category'))
+        {
+            abort(403);
+        }
         $validated = $request->validate([
             'categoryName' => 'required|string',
             'categoryDescription' => 'required|string',
@@ -70,46 +94,75 @@ class AdminController extends Controller
         $categories = $this->CategorieRepository->find($id);
 
         if (!$categories) {
-            abort(404);
+            abort(403);
         }
 
         return view('Admin.Dashboard-categorie-edit', compact('categories'));
     }
 
-    public function updatecategorie(Request $request, $id)
+    public function updatecategorie(Request $request)
     {
+        if(!Auth::user()->HasPermission('edit_category'))
+        {
+            abort(403);
+        }
         $validated = $request->validate([
             'name' => 'required|string',
             'description' => 'required|string',
-            'image' => 'required|image',
+            'image' => 'nullable|image',
         ]);
-
-        $this->CategorieRepository->Update($id, [
+        if($request->hasFile('image'))
+        {
+            $path = $request->image->store('categorie','public');
+        $this->CategorieRepository->Update($request->id, [
             'Nom' => $validated['name'],
             'Description' => $validated['description'],
-            'Photo' => $validated['image'],
+            'Photo' => $path,
         ]);
+    }
+    else
+    {
+        $this->CategorieRepository->Update($request->id, [
+            'Nom' => $validated['name'],
+            'Description' => $validated['description'],
+        ]);
+    }
+
+    return redirect()->back()->with('success','Categorie bien éte mise a jour');
+
 
         return redirect('/admin/categories');
     }
 
     public function DestroyCategorie(Request $request)
     {
+        if(!Auth::user()->HasPermission('delete_category'))
+        {
+            abort(403);
+        }
         $this->CategorieRepository->Delete($request['id']);
         return redirect('/admin/categories');
     }
 
     public function AddService()
     {
+        if(!Auth::user()->HasPermission('manage_services'))
+        {
+            abort(403);
+        }
         $categories = $this->CategorieRepository->ReadCategories();
         return view('Admin.Dashboard-Creation-service', compact('categories'));
     }
 
     public function CreateService(Request $request)
     {
+        if(!Auth::user()->HasPermission('manage_services'))
+        {
+            abort(403);
+        }
         $validated = $request->validate([
-            'name' => 'required|string',
-            'description' => 'required|string',
+            'name' => 'required|string|min:28',
+            'description' => 'required|string|min:28',
             'prix' => 'required',
             'image' => 'required',
             'statut' => 'required',
@@ -140,39 +193,81 @@ class AdminController extends Controller
 
     public function SelectServices()
     {
+        if(!Auth::user()->HasPermission('manage_services'))
+        {
+            abort(403);
+        }
+        $categories = $this->CategorieRepository->ReadCategories();
+        $statistic = $this->ServiceRepository->GetStatisticServices();
         $services = $this->ServiceRepository->ReadServices();
-        return view('Admin.Dashboard-services', compact('services'));
+        return view('Admin.Dashboard-services', compact('services','categories','statistic'));
     }
 
     
-    public function EditService(ServiceEditRequest $request,$id)
+    public function EditService(ServiceEditRequest $request)
     {
-        $services = $this->ServiceRepository->GetServiceByID($id);
+        try
+        {
+        if(!Auth::user()->HasPermission('manage_services'))
+        {
+            abort(403);
+        }
+        $id_admin = Auth::user()->id;
         $categories = $this->CategorieRepository->ReadCategories();
  
             $validated = $request->validated();
 
-            $path = $request->file('image')->store('service', 'public');
-            $id_categorie = $this->CategorieRepository->GetIdByName($request->categorie->id);
+            $id_categorie = $this->CategorieRepository->GetIdByName($request->categorie) ?? 0;
 
-            $this->ServiceRepository->Update($id,[
+
+           if($request->Hasfile('image'))
+           {
+            $path = $request->file('image')->store('service', 'public');
+            dd($this->ServiceRepository->Update($request->id,[
                 'titre' => $validated['name'],
                 'Description' => $validated['description'],
                 'Photo' => $path,
                 'Prix' => $validated['prix'],
                 'categorie_id' => $id_categorie,
-                'prestataire_id' => 17,
+                'prestataire_id' => Auth::user()->id,
                 'created_at' => now(),
                 'updated_at' => now(),
-                'status' => $validated['statut']
+                'status' => $request->statut
+            ]));
+           }
+           else
+           {
+            $this->ServiceRepository->Update($request->id_edit,[
+                'titre' => $validated['name'],
+                'Description' => $validated['description'],
+                'Prix' => $validated['prix'],
+                'categorie_id' => $id_categorie[0]['id'],
+                'prestataire_id' => Auth::user()->id,
+                'updated_at' => now(),
+                'status' => $request->statut
             ]);
-            return view('Admin.Dashboard-Modification-service', compact('services', 'categories'));        
+           }
+           return redirect()->back()->with('success','Le service a été mis à jour avec succès');
+        }
+        catch(\Exception $e)
+        {
+            return redirect()->back()->with('error', 'Une erreur est survenue: ' . $e->getMessage());
+        }
+         
+
+
+          
+            return redirect()->back()->with('success','Le service a été mis à jour avec succès');
 }
 
     public function DeleteService($id)
     {
+        if(!Auth::user()->HasPermission('manage_services'))
+        {
+            abort(403);
+        }
         if (!$id) {
-            abort(404);
+            abort(403);
         }
         $this->ServiceRepository->Delete($id);
         return redirect('/admin/services');
@@ -180,6 +275,10 @@ class AdminController extends Controller
 
     public function unbanService($id)
     {
+        if(!Auth::user()->HasPermission('manage_services'))
+    {
+        abort(403);
+    }
         $this->ServiceRepository->UnbanServiceByID($id);
         return redirect('/admin/services');
     }
@@ -190,14 +289,20 @@ class AdminController extends Controller
         return redirect('/admin/services');
     }
 
-    public function Rendez_vous()
+    public function ReservationsIndex()
     {
-        return view('Admin.Dashboard-rendez-vous');
+        $reservation = $this->ReservationRepository->GetReservationsPaginateAdmin();
+        return view('Admin.Dashboard-rendez-vous',compact('reservation'));
     }
 
 
     public function SelectAvis()
     {
+        
+    if(!Auth::user()->HasPermission('gestion_avis'))
+    {
+        abort(403);
+    }
         $Avis = $this->AvisRepository->ReadAvis();
         // dd($Avis);
 
@@ -215,7 +320,10 @@ class AdminController extends Controller
 
     public function ApproveAvis(Request $request)
     {
-        
+        if(!Auth::user()->HasPermission('gestion_avis'))
+    {
+        abort(403);
+    }
         $this->AvisRepository->ApproveAvis($request['id']);
         return redirect('/admin/avis');
         
@@ -223,6 +331,10 @@ class AdminController extends Controller
 
     public function UpdateAvis(Request $request)
     {
+        if(!Auth::user()->HasPermission('gestion_avis'))
+    {
+        abort(403);
+    }
         $Avis = Avis::with('Prestataire','Client')->findOrfail($request->id);
         if($request->Note)
         {
@@ -247,12 +359,43 @@ class AdminController extends Controller
 
     public function UtilisateurIndex()
     {
+        if(!Auth::user()->HasPermission('manage_users'))
+    {
+        abort(403);
+    }
+        $statisticusers = $this->UtilisateurRepository->GetStatisticUsers();
        $users = $this->UtilisateurRepository->GetAllUsers();
-        return view('Admin.Dashboard-Utilisateurs',compact('users'));
+        return view('Admin.Dashboard-Utilisateurs',compact('users','statisticusers'));
+    }
+
+    public function BanUserByid($id)
+    {
+        if(!Auth::user()->HasPermission('manage_users'))
+    {
+        abort(403);
+    }
+        $ban = $this->UtilisateurRepository->BanUser($id);
+
+        return redirect('/admin/utilisateurs');
+    }
+
+    public function UbanUserByid($id)
+    {
+        if(!Auth::user()->HasPermission('manage_users'))
+    {
+        abort(403);
+    }
+        $uban = $this->UtilisateurRepository->UnbanUser($id);
+
+        return redirect('/admin/utilisateurs');
     }
 
     public function DeleteUser(Request $request)
     {
+        if(!Auth::user()->HasPermission('manage_users'))
+    {
+        abort(403);
+    }
         if($request->id)
         {
             $user = $this->UtilisateurRepository->Delete($request->id);
@@ -263,62 +406,79 @@ class AdminController extends Controller
 
     public function UpdateUser(Request $request)
     {
-        $user = $this->UtilisateurRepository->findUser($request->id);
+        if(!Auth::user()->HasPermission('manage_users'))
+    {
+        abort(403);
+    }
+        $validated = $request->validate([
+            "lastName" => "required|string",
+            "firstName" => "required|string",
+            "email" => "required|string|email",
+            "phone" => "nullable|string|min:5",
+            "image" => "nullable|image",
+            "password" => "nullable|string|min:5",
+            "passwordConfirm" => "nullable|same:password",
+            "Adresse" => "nullable|string|min:5",
+            "city" => "nullable|string",
+            "postalCode" => "nullable|string",
+            "pays" => "nullable|string"
+        ]);
     
-        if ($request->lastName){
-            
-            $validated = $request->validate([
-                "lastName" => "required|string",
-                "firstName" => "required|string",
-                "email" => "required|string|email|unique:utilisateur"
-            ]);
-            $data = [
-                "Prenom" => $request["firstName"],
-                "Nom" => $request["lastName"],
-                "Status" => $request["status"],
-                "updated_at" => now()
-            ];
+        $id_user = $request->id_edit;
+        $user = $this->UtilisateurRepository->find($id_user);
     
-            if ($request["password"] && $request["passwordConfirm"]) {
-                if ($request["password"] === $request["passwordConfirm"]){
-                    $data["Password"] = Hash::make($request["password"]);
-                } else {
-                    return back()->withErrors(['passwordConfirm' => 'Les mots de passe ne correspondent pas.']);
-                }
+        if ($request->email !== $user->Email) {
+            $existuser = $this->UtilisateurRepository->FindByEmail($request->email);
+            if ($existuser) {
+                return redirect()->back()->with('email', 'The email has already been taken.');
             }
-    
-            if ($request->hasFile('image')){
-                $path = $request->file('image')->store('User', 'public');
-                $data["Photo"] = $path;
-            }
-    
-            $this->UtilisateurRepository->UpdateUtilisateur($request->id,$data);
-    
-            if ($request->Role === "prestataire"){
-                $dataPrestataire = [
-                    "zip_code" => $request["postalCode"],
-                    "Numero_Telephone" => $request["phone"],
-                    "Adresse" => $request["Adresse"],
-                    "Ville" => $request["city"]
-                ];
-                $this->PrestataireRepository->update($request->id,$dataPrestataire);
-            } elseif ($request->Role === "client") {
-                $dataClient = [
-                    "pays" => $request["pays"] ?? "",
-                    "telephone" => $request["phone"]
-                ];
-                $this->ClientRepository->update($request->id,$dataClient);
-            }
-    
-            return redirect('/admin/utilisateurs')->with('success','mis à jour avec succès');
-            
         }
     
-        return view('Admin.Dashboard-Modification-utilisateurs', compact('user'));
+        $data = [
+            "Prenom" => $request->firstName,
+            "Nom" => $request->lastName,
+            "Email" => $request->email,
+            "Status" => $request->status,
+            "updated_at" => now()
+        ];
+    
+        if ($request->password) {
+            $data["Password"] = Hash::make($request->password);
+        }
+    
+        if ($request->hasFile('image')) {
+            $path = $request->file('image')->store('User', 'public');
+            $data["Photo"] = $path;
+        }
+    
+        $this->UtilisateurRepository->UpdateUtilisateur($id_user, $data);
+    
+        if ($request->role_id == 1) {
+            $dataPrestataire = [
+                "zip_code" => $request->postalCode,
+                "Numero_Telephone" => $request->phone,
+                "Adresse" => $request->Adresse,
+                "Ville" => $request->city
+            ];
+            $this->PrestataireRepository->update($id_user, $dataPrestataire);
+        } elseif ($request->role_id == 2) {
+            $dataclient = [
+                "pays" => $request->pays ?? "",
+                "telephone" => $request->phone
+            ];
+            $this->ClientRepository->UpdateClient($id_user,$dataclient);
+        }
+    
+        return redirect()->back()->with('success', 'Vos informations ont été mises à jour avec succès.');
     }
+
 
     public function UpdateAdminDetails(Request $request)
     {
+        if(!Auth::user()->HasPermission('edit_information'))
+    {
+        abort(403);
+    }
         $admin_id = Auth::user()->id;
 
             $validated = $request->validate([
@@ -370,7 +530,11 @@ class AdminController extends Controller
             }
         }
     public function UpdatePassword(Request $request)
-    {
+    { 
+        if(!Auth::user()->HasPermission('edit_information'))
+        {
+            abort(403);
+        }
         $admin_id = Auth::user()->id;
             $request->validate([
                 "current_password" => "required|string",
@@ -396,7 +560,7 @@ class AdminController extends Controller
                 {
                     $newpassword = Hash::make($request->new_password);
                     $this->UtilisateurRepository->UpdateUtilisateur($admin_id,[
-                        "Password" => $newpassword
+                    "Password" => $newpassword
                     ]);
 
                     return redirect()->back()->with('passwordchanged','Votre mot de passe a été mis à jour avec succès.');
@@ -405,6 +569,10 @@ class AdminController extends Controller
         }
         public function UpdateImage(Request $request)
         {
+            if(!Auth::user()->HasPermission('edit_information'))
+        {
+            abort(403);
+        }
             $admin_id = Auth::user()->id;
             $request->validate([
                 "profile_image" => "required|image|mimes:jpeg,png,jpg"
@@ -419,8 +587,12 @@ class AdminController extends Controller
 
                 return redirect()->back()->with('imagechanged','Votre photo de profil a été mise à jour avec succès.');
             }
+    }
 
-
+    public function CancelReservation(Request $request, $id)
+    {
+        $this->ReservationRepository->CancelReservation($id);
+        return redirect('/admin/rendez-vous');
     }
